feat: Create new office file in OO
When you create a new "Online office file" (Document, Presentation,
or Spreadsheet) then a modal asks you about document's metadata in a
modal, and when you click on "Create and Edit document" button, the
modal is closed, the item created, and ONLYOFFICE editor is open to edit
the document.

No functional changes for creation of other items.

Part of story #28273: create OnlyOffice documents

Change-Id: Icff26d91e40ba69080093b8f2f35f5d9865f147a