refactor: Extract widgets creation from Angular
part of request #29602 Migrate pullrequest file-diffs to TypeScript

This contribution aims to extract the widget creation part from the
codemirror-helper-service, to migrate it to TypeScript and to have unit
tests for it.

Please note that the widget placement in codemirror part (the one in
side-by-side-diff-component) will be extracted and unit tested in
another contribution.

How to test:
- No functional change expected:
--> The placeholders/inline comments/new comment forms are still displayed properly
--> They work as usual in both diff modes

Change-Id: I4fab7085b2c2d61b9277c6369fc41c6b2bba5a47