refactor: Extract widget management part from diff component
part of request #29602 Migrate pullrequest file-diffs to TypeScript

This contribution aims to extract the part of the
side-by-side-diff-component where we add Comment widgets depending the
line on which they are (added, removed, unmoved left & unmoved right)
and unit test it.

I apologize for the HUGE size of the
side-by-side-code-mirror-widgets-creation-manager test file, but at
least we have unit test for the placeholder height recomputation
mecanism for each kind of line.

In the next step, I'll extract the NewCommentFormWidget insertion part.

Please note that the unidiff mode HAS to keep using the widget_creator
directly, there are no synchronized editors and no state.

[Bonus]

Since comment placeholders creation is ordered by
SideBySideLinesHeightEqualizer through the post_rendering_callback of
InlineCommentWidgets, we can drop
side-by-side-comment-placeholder-builder \o/

How to test:
-- No functional change expected in the side-by-side diff mode
-- No functional change expected in the unidiff mode

Change-Id: Ie0cfab4a9103471b0ffb3397275c6bf2f00edd58