Fix queries to allow to the usage of the ONLY_FULL_GROUP_BY SQL mode
Some parts of docman/document still needs to be fixed but the changes
seems to be a bit more involved than the ones done in this contribution.
They will be managed in a dedicated contribution.

To test you can set the SQL modes in `tests/etc/DBConfig/mysql-server-80.cnf`
and run the tests.

Part of request #29628: Run with ONLY_FULL_GROUP_BY SQL mode

Change-Id: Ib681602dbf2105b6adb9faa7c0f20754e7c59dad