feat: Style ellipsis button
The ellipsis button needs a special treatment: text is white and it has
no border/no background. Furthermore the text has a bigger font size so
that the ellipsis is more visible.

So a new dedicated class `.tlp-button-ellipsis` is introduced.
It has a different look 'n feel when it is used in the header on a
project with a background image.

Part of story #28273: create OnlyOffice documents

Change-Id: I715b4a6b5ab63853e049f165c327ef86de8ef2f5