refactor: Extract code placeholders insertion
part of request #29602 Migrate pullrequest file-diffs to TypeScript

The insertion of code placeholder is now delegated to the
SideBySideCodePlaceholderCreationManager (formerly
SideBySideCodePlaceholderBuilder). Since this object was creating code
placeholder widget creation parameters, it was the best candidate to
also command the insertion of those placeholders when they are needed.

[Bonus]

I've fixed a bug that was occuring on the first line of the file, when
this one was added. Now the opposite line should have a code placeholder
above.

How to test:
-- No functional change in side-by-side diff mode

Change-Id: Ibf6ece438efcea0e3047dac3f1c9cadb24cf40d9