Fixes request #29631: REST endpoint artifacts/:id/changesets is slow when many changesets exist
This contribution changes the way the result is built: instead of
retrieving and building data for all the changesets of an artifact and
then keep only what we need for the page we are currently looking at, we
first keep only what we need and then we do the heavy work.

Note this can be even better by only retrieving the changesets we need
first but at this stage this is unlikely to be an issue. Retrieving and
instatiating the high level Tracker_Artifact_Changeset is fairly cheap
in comparison of retrieving all the information associated with it.
If/when retrieving all the changesets of an artifact becomes an issue it
will require far more work and re-design because it means that
`Artifact::getChangesets` has to go.

Change-Id: I6d9d173f72c5a905f3863cc7f1b86aa46e9abbab