feat: Scroll to file-diff comment
story #29256 have a link to comment in context from overview

This contribution allows user to go directly on a given comment from the
timeline view when they click on the link in the comment's header.

Please note that the way we scroll the comment widget into view is
kind of tricky. Indeed, we have to call cm.scrollIntoView first to the
target line number (the one having the widget), then refresh the editor,
and scroll the widget into view afterwards. The whole process is wrapped
in a setTimeout callback in order to let the comment being rendered.

It was quite painful to make it work because codemirror only renders a
subset of the file for performances reasons. If we have placeholders and
comments in the file, we have to scroll a first time to the line, then
refresh the editor, so the widgets have been rendered, then scroll the
widget into view, again.

Please also note that in some cases, the scroll is not precise enough,
and the user might need to scroll a bit to see the comment.

I do not have any alternative solution to propose, and this one seems to
work 99% of the time, so I guess it is goodenough ¯\_(ツ)_/¯

How to test:
- In the timeline view, click on the link in the header of an inline
  comment
--> You are redirected to the file diff view on the given comment
--> Same expectations for both diff modes

Change-Id: I181602a5aab44e2d49c877afb1bd470519595f3b