feat: Duplicate permissions at project creation
Given a project with mediawiki standalone activated,
And READ permission is given to project members and one of static
ugroups,
When a project is created from this template,
Then permissions are duplicated.

It has been decided to make this duplication synchronous with project
creation:
* Create a dedicated async task for this will block us in the future if we
want to make duplicate-permissions-task and
create-mediawiki-instance-task dependent of each other.
* Existing create-mediawiki-instance-task is only launched when service
is activated, this means that if the service is not activated in the
template, then the permissions will not be inherited and they will
default to project members, which is not necessarily wanted by template
maker.

Part of story #25733: manage mediawiki permissions

Change-Id: I873d5776e099a15eace33718762773557237c230