fix: Fault notification clearer clears too many times
Part of story #24971 [modal] search for artifacts

The fault notification is cleared whenever something happened with the
autocomplete, the consequence is the user cannot see the error feedback
in the modal.
With the fix, the user should be able to see the error feedback while he
is not typing something in the autocompleter.

How to test:
 - I suggest to add a `throw new RestException()` in the
   ArtifactsRessource.class.php L419
 - Display the artifact modal
=> Nothing is displayed.
 - Open the artifact link selectbox
 - Type some artifact id
=> The feedback is displayed
 - Type some artifact id
=> During the typing the feedback error is cleared
=> At finnish typing, the error feedback is displayed
 - Clear the autocompleter
=> the error feedback is cleared

Change-Id: I289ddc92756774d0dfc49ce3b186e8eeb0cbe046