fix: do not receive emails even when in global notifications
Initial implementation was only dealing with notification based on user
relationship with the current artifact (submitted, assignee, etc).

When in global notification, the user is actually referenced by its
email rather than by its username. This commit clears both.

In order to test this patch you should have a user that is in global
notification of a tracker (either as user or in group) and this user
should have unsubscribe to the notification made by its own changes in
preferences. With this patch, they should not receive any email.

Part of: story #10469 choose if I receive a notification when I create/update something

Change-Id: I6c528b2a9677262e1668964346ad69cc8ca9bd0b