feat: call search endpoint
How to test:
- Enable the feature flag:
$ tuleap config-set feature_flag_history_and_search_in_modal 1
- Install a full text search backend and index your artifacts
- Open an artifact (with an artifact link field) in the artifact modal
- When you choose any link type except for "Parent", when you type one
or two characters, the list of recently viewed items is filtered
- when you type 3 or more characters, a "Search results" section should
appear. If your query matches artifacts, they are shown in the dropdown.
Otherwise, you get an empty section.
- when you switch link type or empty the search field, the section
disappears from the dropdown.

Note: errors are not handled yet, they will be handled on a next
contribution.

part of story #24971 [modal] search for artifacts

Change-Id: I4fab5a5fbd89ce9fdae00c3a6f63d4a109b52f70