feat(forge-config): Docker image should have a way to declare the exposed SSH port
This contribution introduce a change in the way environment variables
are managed by PHP FPM. Until now we were using the default config where
environment is reset by PHP FPM, it's no longer the case. It's needed so
the PHP code run in front end have access to the variables defined with
Environment Variables.

Previous work made workaround by writing the values in `local.inc` or
`database.inc` but this cannot be done for plugins and re-writing files
at Docker image start feels not very convenient to do (esp when not all
variables are transfered like in git with this contrib).

Closes request #29649 Docker image should have a way to declare the exposed SSH port

Change-Id: I296adec35e4ced7d8101725d6a3360f1903ae9ee