chore: Do not allow planning update with backlog trackers that have hierarchy links
This is part of request #29972 Planning backlog trackers must not have hierarchical links

When editing the backlog trackers of a planning, we must not be able to select trackers
that have a hierarchical link, because this will mean that both parent and child artifacts
can be planned in different milestones.

Note: A more helpful error message with trackers information will be added in a next commit.

How to test
-----------

* Have a project with the agiledashboard service enabled
* Define a parent/child relationship between a tracker A and a tracker B
* Go to your agiledashboard administration, and select a planning
* Edit the backlog trackers of this plannig by selecting both trackers A
  and B

=> The backlog trackers must not be updated

Change-Id: Iff84b674a8966beda1bf09b9cdfaa88c3f074dae