MediaWiki plugin does not need to have the MediaWiki Standalone plugin activated
The only need of the MediaWiki plugin is to access a class located in
the MediaWiki Standalone plugin. We aensure the files are present on the
filesystem via the RPM dependencies.

Force activating the plugin via a forge upgrade is not adequat as it
bypasses some the actions done during the activation (e.g. creation the
`/etc/tuleap/plugins/<plugin>` folder). A proper way to manage that
would be for the upgrade mecanism to check if all the dependencies are
correct before doing anything else and force activate the missing
plugins.

Part of story #25733: manage mediawiki permissions

Change-Id: I4d94d419a66b0cdbb2ef7d2d8b533c7bdd474170