fix: Avoid ResizeObserver loop error
How to test:
- end-to-end tests for the Artifact modal should succeed.

Notes:
When you change the size of something during a ResizeObserver callback,
it can cause an error because the observer could not report all
observations during a single animation frame [0]. Wrapping in
`window.requestAnimationFrame()` callback [1] solves this problem: the
change of the dropdown position and size comes later than the
ResizeObserver callback. It's also a good idea performance-wise.

part of story #24971 [modal] search for artifacts

[0] https://stackoverflow.com/a/50387233
[1] https://stackoverflow.com/a/58701523

Change-Id: I4895710058beee27ce18fe8333c809e82234285c