chore: Close request #29973 User mail in Git log viewer of Tuleap is case sensitive
When I display the commits in a Tuleap Git repository,
if the email set in my git commit is not strictly the same as the one defined
on my Tuleap account (with the same case), the commit is not linked to me.

Emails are in a vast majority case insensitive, this must be taken into account.

How to test
-----------

* Clone a Tuleap Git repository and add a commit with the commit email
  address the one used by your Tuleap user with a different case
* Push this commit

=> While browsing the commits list in the repository, this commit must
be linked to your Tuleap profile.

Change-Id: Ied1c447e363d5d00d7f9bb207fe1819236194728