refactor: Make user and current PR parameters of the controller
No functional change expected when creating replies to inline comments
in overview, unidiff or side-by-side diff views.

Since those properties are never useful to the template part, we can
move them to the controller. It avoids some back-and-forth, as
previously we had to init the custom element just so it could pass them
back to the controller.

part of request #29602 Migrate pullrequest file-diffs to TypeScript

Change-Id: I1ded18a5e9ff1691b2f11cd83627ac650f92e4c2