chore: Have a more meaningful error message for backlog trackers error
This is part of request #29972 Planning backlog trackers must not have hierarchical links

When triggering the error by selecting both parent and child trackers as
backlog trackers of a planning, then the error message must have a more
meaningful content to understand where is the error.

How to test
-----------

* Have a project with the agiledashboard service enabled
* Define a parent/child relationship between a tracker A and a tracker B
* Go to your agiledashboard administration, and select a planning
* Edit the backlog trackers of this plannig by selecting both trackers A
  and B

=> The backlog trackers must not be updated and the error message must
contains the trackers that cannot be selected at the same time.

Change-Id: Ibb526a171487af7f3d897f033bed5004e0ddcd30