feat: Introduce table to manage OO servers
Site administrators are supposed to be able to manage more than one
server. This contribution changes the way we display the current default
server to prepare the field.

* If nothing is configured, the table is empty and user can add a new
server
* If there is an existing server, the table has one entry and user
cannot add a new one (for now), but can edit it.

This is only a frontend change, backend is left as is.

Part of story #29981: restrict an OnlyOffice server to some projects only

Change-Id: I8b3779812e28a9f69747d4e26b660750f94eb4f6