chore: Close request #29987 Massmail notifications can be send one by one
Currently, when sending massmail notifications, they are sent for a batch of 25 users.
If sending this notification fails for one user, then the 24 others users
will also not receiving it.

We have to avoid it and give the ability to send these massmail notifications one by one.
To try this new behavior, we have to activate a feature flag.

How to test
-----------

* Enable the feature flag tuleap config-set feature_flag_send_massmail_one_by_one 1
* Send a massmail

=> Each user involved in massmail must receive its notification
individually

Change-Id: I66bb44f0b6d0fd8329907da2537cae5ea0a4a3a7