refactor: do not store anymore settings in forgeconfig
No functional changes. Now server settings are stored in the dedicated
table.

For now we take the first available server for communication between
Tuleap and OO document server, but a next contribution will tie a
download/save token to a server so that we know which JWT secret key we
can use to ensure authenticity of the JWT.

Part of story #29981: restrict an OnlyOffice server to some projects only

Change-Id: I06bb96a231fc43915d7427c0532beba1e6db1272