refactor: Event architecture for field communication
How to test:
- Open the artifact modal (create or edit).
- If the tracker has at least one file field, you should see the "quota
display" at the bottom-left corner.
- If the tracker has no file field, it should not appear.

Notes:
It might seem overkill "just" for this use-case, but it's a pretext to
introduce the Event dispatcher. Events will be useful for field
dependencies of list fields, for file uploads before submit, for submit
(no need of another big switch-case, each field is responsible for
formatting its values into a REST payload).

part of request #22648 Convert AngularJS directives and controllers to
hybrids web components

Change-Id: Ibec4857b227f10efbf6d99978ce25aa042892a2f