feat(el9) add support of Subversion
ViewVC is no longer available in epel because the last version of ViewVC
(1.2.x) is python2 only and there is only python3 on RHEL9.

This contribution builds ViewVC from sources / master branch is it
appears to be ready for primetime except a feature that we don't use
related to CVS db: https://github.com/viewvc/viewvc/issues/250

At the moment, the only thing that seems missing is Pygments so there is
no syntax highlighting. I'm not sure how this should be handled properly
so it's left for a future patch. FTR, python3-pygments is in `devel`
rocky repository.

The Tuleap built ViewVC is built on both el7 and el9 because we don't
have a good mean to distinguish the 2 builds right now. It's easier to
build for both and to prevent viewvc-tuleap (the one for el9) to be
installed on el7 with python3 and subversion 1.14 requirement.

Finally, Tuleap code requires en_US.UTF8 lang to works properly but it's
no longer part of the default installation, hence the new requirement.
In the same spirit, httpd is no longer pulled by mod_dav_svn so we need
to add the explicit dependency.

Part of: request #29619 Add support of Rocky Linux 9

Change-Id: If72de0ea0b305e98fd88392b8bb29bff54dc9d40