refactor: tie save token to a specific server
Follow-up to git #tuleap/stable/903399279500ac84c08646a4c34467d89dbf49f8

No everytime a user open a document in OO, the save token is linked to a
server_id. This will allow later (when we will have more than one sever)
to use the right key to verify the JWT.

In order to not invalidate existing sessions during the platform
upgrade, their token is not linked to a given server and in this case we
will take the first server like before. This behavior should be removed
after a bunch of releases to let time to admin to upgrade their
platform.

Part of story #29981: restrict an OnlyOffice server to some projects only

Change-Id: Ibfb5072a39d59a52ee2d80fb94158b5d59f1f1bc