Disable Vitest threads when collecting coverage
It might help with the tests hanging (sometimes) when
collecting coverage. If the issue still manifest we should
try to use Istanbul instead of c8 with Vitest.

Note that running Vitest without the threads enabled break down some
isolation. It should not really be an issue since most of our runs are
done with the threads enabled.

part of request #27556: Migrate Vite-based packages to Vitest

Change-Id: I4011773208d463a8a3715cbfc0d3fd54d9a8cbef