chore: Initiate fake pagniation for Baseline
This is part of request #29995 Enhance handling of pagination for Baselin plugin

The very first step is to create a fake pagination (handled by the JS
code and not the REST API directly) to allow to process more than 100
linked artifacts.

Note: the rework of REST API pagination handling seems quite huge, so I decided to do baby
steps to the final result.

How to test
-----------

* Have a release with lot of artifacts and children (more than 100
  links)
* Create a baseline

=> No error must be here while displaying the baseline.

Another way to test this patch is change the ```limit``` value in
Javascript code, rebuild the app and check that everything is well
displayed.

Change-Id: Ia5661d400ce1e5cf9e2b1fa0856a08ef8aad17ba