fix(el9): update rockylinux:9 image preparation
It seems that some changes were done on systemd in base image so our
clean-up must adapt. In order to be in control of those changes, the
image is now pinned. The pin is done on the current image.

This commit also change the way distlp tests are run to gather
informations about failure in Nightly.

Part of: request #29619 Add support of Rocky Linux 9

Change-Id: I5641bfd8f5fc9a2291d356e56602f8db9d163197