refactor: Use cy.session to log in AgileDashboard tests
No functional change.
End-to-end tests for AgileDashboard should succeed.

Notes:
This introduces new commands to make migration from old ones easier. It
introduces confusion to have two ways to log-in, but it lets us work on
each plugin gradually instead of a big bang migration. At the end of the
work, we should delete the old login commands and rename the new ones.

part of request #29992 Bump cypress to 12.1.0

Change-Id: Ib4490e469b7273d9ce9ea03872f7d2b26f5de1dd