Closes request #30001: Reject SVN requests without Authorization header as soon as possible
No functional changes expected but it should save some resources.

When you checkout a SVN repository you should not see the following log
line since the request will be rejected directly by nginx:
```
[debug] Received SVN access request is incorrectly formatted, no credentials found
```

This avoid reaching out the PHP backend for something we know in advance
that it will be rejected.

Change-Id: Iddeb14f72afa163626d9c6ac89ce1203eb59fd16