fix: Close request #30009 Selecting a new renderer type wrongly closes the dropdown
When adding a new renderer in a tracker report, we can select a renderer type (table, cardwall, graphs).
When selecting a new value for the renderer type in the dropdown,
the dropdown will be closed and have to be removed to finish the renderer creation.

This occurs only on Chromium based browsers. All works as expected with Firefox.

How to test
-----------

* Go to a tracker you are admin with a Chromium based browser
* Click on the "+" button to add a new renderer
* Select a new renderer type

=> The dropdown is no more closed (must continue to work as expected on
Firefox)

Change-Id: Id752a21625a1aa72ed1766fff40e77e16de631b6