feat: autocompleter for project to allow
Go to site admin » ONLYOFFICE

Have a server that is restricted (you may have to hack the database).
You should be able to search for a project to be allowed. The project
cannot be added yet, the corresponding button is disabled for now.

Allowing a project for a server is intertwined with its existing
presence in other servers (a project cannot be allowed for more than one
server). In order to anticipate this exclusion (and the fact that
project autocompleter developer experience is poor), the frontend is now
a Vue3 app to have fine control of various components.

No functional changes in displaying/adding/editing/removing servers.

Part of story #29981: restrict an OnlyOffice server to some projects only

Change-Id: I6c8bd50f2e9a1f75707b9fd2c012fefdb281b0d8