Fixes request #30015: Users should not be logout from MW standalone instances when they authenticate on another service
Issue only occurs on instances where LDAP is used because the LDAP
plugin triggers an update after every succesful auth.

This contribution improves this behavior by triggering an update only
when it is absolutly necessary.

To test try to authenticate against a SVN repository using your
password, it should not trigger an deauth event for the user on MW
instances (the easiest way to see it, is to have the logs at the debug
level to see the events).

Change-Id: Ife8514acc0fd36a6504278710d3d3f339bedbe85