feat: Add (virtually) project in the allowed projects
Go to siteadmin » ONLYOFFICE

Open the restrict modal of a retricted server.

Search for a project and click on `Allow project` button: the project
should be added in the table. If an error occurred during the fetch of
project information, then an error message is displayed.

We are forced to make a REST call because project autocompleter result
is a string "public name (shortname)" and we have to retrieve other
information of the project.

The added projects are not saved in the database, it will be part of a
dedicated contribution.

Part of story #29981: restrict an OnlyOffice server to some projects only

Change-Id: I057285f9ce414ee7991e7e239b7023995e1ee93d