feat: Be able to reopen an abandoned pull request with REST
This is part of request #30013 Be able to reopen an abandoned pull request

Given a pull request that have been abandoned, it's currently not possible to reopen it.
User have to create another branch and create a new pull request.

Tuleap must offer the ability to reopen an abandoned pull request.
This first step is to be able to do that using the REST API route
PATCH /pull_requests/:id with the current payload:

{
   "status": "review"
}

How to test
-----------

* Create a pull request asking to integrate branch B into branch A
* Abandon this pull request
* try to reopen this pull request using the REST route

=> The pull request must be reopen.

Change-Id: Ic4313ca6860020ec23afd196110985c6fa4cc707