feat: Display indication on which project is being allowed
User should have a clear view of which project is being added, and which
project is being removed from the allowed list in site admin »
ONLYOFFICE » server » restrict.

Success/danger rows styles are added in TLP so that we don't have to
repeat ourselves (e.g artifact link selector).

Part of story #29981: restrict an OnlyOffice server to some projects only

Change-Id: Ia118b82f984b1b101022b79133142dcad1e7710b