fix request #30024 DynamicCredentials prevent svn caching mecanism to work
SVN auth caching mecanism relies on nginx fastcgi_cache. However this
cache is not used when the upstream (SVNProjectAccessController here)
set a cookie. Under regular conditions this is the case, however when
dynamic credentials plugin is used, it auto start sessions as part of
beforeLogin process. This set a cookie and defeat the caching mecanism.

This patch walk around all PHP_Session::start() usage to prevent the
start of session when the environment explicitly tells not to.

Change-Id: Ic231684ecd9cafc3c1a0d3ee12ac78931eae79be