chore: Build iterations app separately
No functional change expected in Iterations app. Given a Program project
with PIs and Iterations children of a PI, when you browse the PI in
Program management, click on one of the PIs and click on the "Plan
iterations" link. The app should load correctly, and it should be
translated. Styles should be loaded.

part of request #29650 Split program management frontend into separate
apps

Change-Id: Id833aed9ca76e727fc27cb430c0952cde3e08e5c