feat: Save allowed/revoked projects
Now sitadmin is able to allow/revoke projects to use a given document
server.

If a project is allowed for another server, then it will be removed
before being added to the new one. Siteadmin is warned about this during
project selection.

Part of story #29981: restrict an OnlyOffice server to some projects only

Change-Id: I39797bbb4115dc34caa62292269647d2075f1dbc