fix: DynamicCredentials (non) session storage
This is a follow-up commit of 4da4a83de0fc48d4bd12168f7cdab61b3f80e60f

The original fix completly discarded Dynamic Credential usage in the
context without session (SVNAuthController ATM) but it was not perfect.

This patch uses a regular array for storage (instead of session) when
session are not allowed to persist "session" information. The array is
static to be persisted across the request (a Singleton might have been a
more explicit solution but was less convenient to implement).

Part of request #30024 DynamicCredentials prevent svn caching mecanism to work

Change-Id: I35073cd3f1bf5770d61fbc4733c41ff20e43d8f6