chore(distlp) Run SVN CLI tests as part of the main test suite
SVN CLI tests are now run by full e2e test suite.

In order to avoid a mess with Cypress docker image, the choice we to
have a dedicated phpunit test suite (it's a copy of the tests that were
run on distlp architecture). Cypress is designed to run browser tests,
using it as a general purpose runner would have been a long term
maintenance nightmare.

Finally, in order to have a working subversion environment for Tuleap I
moved the code that was in the legacy setup.el7.sh into a Tuleap
Command. As part of this work, I clean-up a bit the apache config. The
include line is removed because the setup was actually removing it so it
was never deployed with this line. Same for the other transformations.
Instead of sed/awk a file we manage end-to-end, it's easier to modify
the source.

The choice of having it as a Tuleap CLI command (and not a TuleapCfg
command) is that we want to have it at plugin level and it's not
possible to do that in TuleapCfg and in all cases, this will be run
after the initial Tuleap configuration so we don't have to bother with
the availability of the database (main concern and reason to exist of
TuleapCfg). As a side effect, when the rpm package is installed, the
plugin will always be configured and Tuleap installed/activated
(previously it was only configured).

Part of: request #29998 Remove Distributed Tuleap (distlp) support

Change-Id: I9c5b1b4ba964459a76bcfe0893660e06449ad4a6