fix: display feedback errors on user registration form
When the register form is invalid (e.g: the email is not submitted),
then a feedback was displayed until 89dedaf5222f074593f980bb77aa2938d30ea203

This was due to the fact that we were in a mixed state
BurningParrot/FlamingParrot. Now that every pages of register have been
converted to BurningParrot we can now remove our custom hack. For the
time being we insert `register.php` in legacy
BurningParrotCompatiblePageDetector but it will be removed when
dedicated controllers will be extracted (next contributions).

No functional changes in confirmation pages.

Part of request #29605: Convert account registration pages to BurningParrot

Change-Id: I055e19c86dfec9d2623f1be0a7a22b010262ddd1