feat: Close request #30354 Copied artifact must be set in explicit backlog if source artifact is part of planning
Given an artifact that is set in the explicit top backlog of the project,
or is planned in any milestone of the project. If this artifact is copied,
then the copied artifact must be added in the explicit top backlog of the project.
The same goes for the copied children if there are any.

How to test
-----------

* Have an artifact either planned in a milestone or set in the explicit
  top backlog of a project
* Copy this artifact

=> The copied artifact must also be added in the explicit top backlog of
the project

Change-Id: I34d487f896fd069fef55726c11078f87efb28af4