feat: list-picker also watches changes of value attribute
No functional change expected.

How to test:
- In TLP doc, go to the list-picker page. Open your dev tools, find a
<select> element with a list picker attached.
- Change the text content of one of the <option> elements. Nothing
happens.
- Change the value of the same <option>. List-picker should refresh and
change the contents of its dropdown with the up-to-date label and value
for the option.

Notes:
This change is made in order to handle <option> elements rendered by
hybrids. Hybrids tries to minimize the changes it applies to the DOM, so
it only updates the text content and attributes of options when they
change. Since the actual <option> node does not change, list-picker
became disconnected and allowed selecting options that were no longer
present. It is surprising that we did not run into this problem earlier,
Vue is supposed to have a virtual DOM and also minimize changes, but
apparently it recreates new <option> elements.

part of request #22648 Convert AngularJS directives and controllers to
hybrids web components

Change-Id: I8f31c9f0c18cff7b5ff9583ca8df10f29da113a8