Refactor: convert restrict modal to dedicated view
We need to switch to a full featured page instead of a modal for future
developments, mostly due to the fact that we need to redirect to the same
page, and preopening a modal is not the better way to do that.

To test:

Click on [restrict] button leads to a dedicated page.
User can refresh the page.
User can make changes to restrictions and submit. They are redirected to
the same page.
Click on [cancel] button leads to servers list.

Implementation note: instead of using a dedicated routing library like
Vue Router, we start small by using our own since our needs are quite
limited. If needed we will switch later to Vue Router.

Part of story #29981: restrict an OnlyOffice server to some projects only

Change-Id: Ibaf8dba12843aa1013f5d0a2743b9a5a1cff4285